third_party/WebKit/LayoutTests/media/encrypted-media/encrypted-media-needkey.html

<!DOCTYPE html>
<html>
    <head>
        <title>NeedKey</title>
        <script src="encrypted-media-utils.js"></script>
        <script src="../../resources/testharness.js"></script>
        <script src="../../resources/testharnessreport.js"></script>
    </head>
    <body>
        <video id="testVideo" controls></video>
        <div id="log"></div>
        <p>Test that needkey event is fired on an encrypted media file.</p>
        <script>
            var expectedInitData = stringToUint8Array('0123456789012345');

            // Will get 2 identical events, one for audio, one for video.
            var expectedEvents = 2;

            async_test(function(test)
            {
                var video = document.getElementById('testVideo');

                var onNeedKey = function(event)
                {
                    assert_equals(event.target, video);
                    assert_true(event instanceof window.MediaKeyNeededEvent);
                    assert_equals(event.type, 'needkey');
                    assert_array_equals(event.initData, expectedInitData);

                    if (--expectedEvents == 0)
                        test.done();
                };

                waitForEventAndRunStep('needkey', video, onNeedKey, test);
                video.src = '../content/test-encrypted.webm';
            }, 'Needkey fired on encrypted media file.');
        </script>
    </body>
</html>
