third_party/WebKit/LayoutTests/media/encrypted-media/encrypted-media-needkey-expected.txt

Test that needkey event is fired on an encrypted media file.

EVENT(needkey)
EVENT(needkey)
This is a testharness.js-based test.
PASS Needkey fired on encrypted media file. 
Harness: the test ran to completion.
